move pbkdf2 settings to test_settings.py file
